Mas cambios mas rest

+++ src/main/java/com/lmh/repository/ISalaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.lmh.entity.Sala;
 
 @Repository
-public interface ISalaRepository  extends IBaseRepository<Sala, Long>, QuerydslPredicateExecutor<Sala> {
+public interface ISalaRepository  extends IBaseRepository<Sala, Integer>, QuerydslPredicateExecutor<Sala> {
 	
 }
